requirements.txt
tensorflow~=2.19.0
numpy~=2.2.6
opencv-python~=4.11.0.86
pillow~=11.2.1